+++ src/azure-cli/HISTORY.rst
@@ -20,7 +20,7 @@ Release History
 
 **Cosmos DB**
 
-* Add `--type <key-type>` to `az cosmosdb keys list` command to show  key, read only keys or connection strings
+* Add `--type <key-type>` to `az cosmosdb keys list` command to show key, read only keys or connection strings
 * Add `regenerate` to `az cosmosdb keys` group
 * Deprecate `az cosmosdb list-connection-strings`, `az cosmosdb regenerate-key` and `az cosmosdb list-read-only-keys`